+++ Assets/Scripts/UserInput/InputHandler.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: ae73d10ed03bccf4a87a5ea85f35d753
-timeCreated: 1522400480
+guid: b045f41a57beccd4d85245b95be5d10a
+timeCreated: 1522400086
 licenseType: Free
 MonoImporter:
   externalObjects: {}
